Front End of the Home Feed
Note, this can all potentially be modified
--Necessary Stuff--
MainActivity.java - Code for giving cardView swiping behavior. It contians some temporary values for testing.
arrayAdapter.java - Code for handling and passing imformation to the cardView. It uses the temporary values in MainActivity to populate the card deck. The method populationLogic() that gets that information will likely need to be changed.
Property.java -  defines the property object that will be used to store values from the server so that our methods can pass this information easily.
activity_main.xml - the main portion of the UI
property_card - defines the layout of the card and is placed into the UI defined by  activity_main
circle.png & fade.png - resources being used by the xml files
(be careful of other small changes, such as the new style added in styles.xml that is refecned in the UI xml files)
--Not Necessary Stuff--
ViewProperty Class - just a temporary activity that can be replced by the "addingProperty" activity. Note its existance is documented in the manifest file.
view_property.xml - temporary UI